query-service/pyproject.toml
[tool.poetry]
name = "query-service"
# LLM_REFACTOR_FINAL: Set final version number
version = "0.3.0"
description = "Query service for SaaS B2B using Clean Architecture, PyMilvus, Haystack & Advanced RAG"
authors = ["Nyro <dev@nyro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tenacity = "^8.2.3"
structlog = "^24.1.0"

# --- Haystack Dependencies ---
haystack-ai = "^2.0.1" # Core Haystack (Document, PromptBuilder)
pymilvus = "^2.4.1"    # Milvus client
# FastEmbed integration for Haystack (still used for embedding)
fastembed-haystack = "^1.4.1"

# --- LLM Dependency ---
google-generativeai = "^0.5.4" # Gemini client

# --- RAG Component Dependencies ---
sentence-transformers = "^2.7.0" # For BGE Reranker adapter
bm2s = "^0.2.0"                # For BM25s sparse retriever adapter
numpy = "^1.26.4"              # Often a dependency for ML/vector libraries

# Optional for performance
# onnxruntime = { version = "^1.17.1", optional = true }

[tool.poetry.extras]
onnx = ["onnxruntime"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"